// src/addres.dto.ts
export interface AddressDTO {
  name: string;

  szam_country: string;
  szam_postalCode: string;
  szam_city: string;
  szam_streetAndNumber: string;


  szal_country: string;
  szal_postalCode: string;
  szal_city: string;
  szal_streetAndNumber: string;

  couponCode: string;

  cardNumber: string;
  expirationDate: string;
  securityCode: number;

}